interdb script works
Allows for creating two databases; one with any name and one called "eenmaalandermaal". idb_eenmaalandermaal needs to be executed in "eenmaalandermaal", all other scripts need to be run in the other database with the tables from the databatch. All inserts will be automatically converted to fit in the correct tables.